backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
social-auth-app-django
mysqlclient
pymysql
requests
python-dotenv
PyJWT
google-generativeai
tiktoken
python-jose
openai
anthropic
gunicorn
whitenoise
dj-database-url
psycopg2-binary